fix: ignore Snyk ws@1.0.1 patch for now
engine.io-client@1.6.11 depends on ws@1.0.1, which carries a security
vulnerability. Snyk provides a patch for the vulnerability, but it
brings the behaviour of ws@1.0.1 in-line with ws@1.1.1, which
engine.io-client is not compatible with.

Snyk is currently working on an updated patch that is compatible with
this version of engine.io-client. In the meantime, we're going to have
to stay vulnerable as we depend on the impacted functionality.