feat: rewrite raw endpoints to /contents api for GH
Since GH enterprise 2.11 the `/raw` endpoint is not supported.
In order for us to make a gradual change without breaking our users broker clients
I added logic that rewrites /raw to /repos/owner/repo/contents endpoint.
Once all of our GHE users will use the we could change all our system code that touches GH files, or hopefully will do it on the GH agent, and delete this code.

I know that this code makes our filter logic GH aware, and this was not the intention of the broker, but this was the solution for non breaking change.