fix: consider the case where body is 'null'
when getting the body it could be a string representation of null, meaning 'null'...
in this case Buffer.from('null') will return 'null' and JSON.parse will parse it as null.
eventually this chain of event will cause tryJsonParse to return null and relay.js will fail on line 114
because parsedBody will be null.