fix: do not decode urls when proxying them
We used to decode urls to 'help' them match the approve filters. This is misguided as the broker is not supposed to 'mangle' the requests it is proxying on top of what is defined in the approve filters.

This is breaking some functionality when the proxied service expects an encoded URL, but the broker decodes it.

Applying this change will require adjustment of our pre-defined approve filters. For example, approving the fetching of `package.json` files in any depth of a git repo will require approving `.../package.json`, `...%2Fpackage.json`, `...%252Fpackage.json` and so on. The needed changes to the pre-defined approve filters will follow on a separate commit.