Update accept.json.sample
Please credit THG if change is accepted.

Fix for Artifactory CR integration. Snyk is not calling the v2/token endpoint and using the Token for auth. As a result, you need to add the auth basic scheme to every v2/* call otherwise you will get 401 unauthorised.

Tested that adding this resolves the issue with authentication.