Migrate tests to Jest
We're in the process of moving the Broker to TypeScript so that we can
improve our confidence when upgrading dependencies and making other
changes. However, the legacy testing frameworks and tools were making
progress slow. Given the tests need to also roll forward, we've opted to
migrate these first. In general, the tests should be mostly identical,
but instead of using `proxyquire`, `tap`, `sinon` and `require-reload`
we only use Jest.

The aim is not to make the tests "better" yet, only to focus on
modernising the infrastructure without any changes to the source code.
This should make the test migrations easier to read and review.

- Delete the object assign test. I trust Node not to break
  Object.assign()
- Re-enabled the proxying-decision test. This had a bad file extension,
  so it probably wasn't picked up by Tap for the past 3 years