fix: replace partial chunk
Fix bug replacing 'RES_BODY_URL_SUB' in relay.
The function returns an object with newChunk.
This should have been assigned to responseBody instead of the entire
object.

The effect of this was to transform our responses from a string to an object,
which when we attempt to write to the http response it throws 'ERR_INVALID_ARG_TYPE'.

Added 'RES_BODY_URL_SUB' env var to server-client functional test to
assert desired behaviour.