fix: Change documentation language referencing Bitbucket Server to Bitbucket Data Center.
This PR is a simple documentation update. Bitbucket Server is now a
deprecated service and Atlassian is moving to Bitbucket Data Center
which is essentially the second iteration of BB Server.

Broker should continue to work as-is but the documentation could be
confusing for those looking specifically for Bitbucket Data Center
support.