fix: set permissions for OpenShift
OpenShift has very specific needs for permissions - all files should be
part of the "root" group, and should have the same permissions as the
primary user.

This needs to be done per-dockerfile so that can adjust permissions for
all files, including those generated during the build.

Also remove global permissions from all files and folders while we're
doing this.