feat: reject large bodies in the client
Increase the maximum allowed body size to 21MB on the server side, so we
can actually accept files of up to 20MB, and reject requests on the
client side that have a Content-Length of over 20MB.

It's better to do this on the client side as the server will close the
WebSocket connection if too much data is passed through it. This is
quite disruptive and leaves the original HTTP request hanging until it
times out.

This also sets the .nvmrc to match the actual Node version used, and
fixes a typo in the README.