Drop Ubuntu version to 20.04
Ubuntu 22.04 has some issues with older versions of the Docker Engine
in Kubernetes. It's not clear whether this is a bug in Ubuntu / the
Ubuntu docker image, or if it's a bug in the Docker Engine. Either way,
in order to retain support for older versions it is necessary to drop
the Ubuntu version back down.

Symptoms of this include segfaults or hard-crashes with C/C++ stack traces.