fix: remove callback that was causing memory leak
Removed a callback that would handle the case where a response was emitted to a streaming request, which isn't supported. Unfortunately, this resulted in a substantial memory leak because the callback was never removed. This only affected the Broker Server, and then only if a large number of streaming requests were received.

Also added better handling for errors - e.g., destroying the stream to trigger proper error processing, and adding content-length/content-type for errors POST'd to the Server.