fix: Server was encountering an error when streaming was requested from Client
Streaming requests received from the Client were not being handled correctly (the server was attempting to POST the response to the client), which was causing the Server to crash. In general, this seems like a configuration error on the Client side, but it is happening in production so needed to be fixed.

This has been fixed and responses will be correctly sent to the Client.