Add environment variable to strip all X-Forwarded-* headers
As some backend systems (e.g., Nexus) use these headers in unexpected ways, it is safer for us to strip out any header starting with X-Forwarded-, along with the Forwarded header.

Configurable via the REMOVE_X_FORWARDED_HEADERS, so it can be turned off if there are any unexpected issues.