Add Dispatcher calls to Server
When the server starts up, it should notify the Dispatcher that a new process is available to handle requests, and when it shuts down it should notify it that it is no longer available.

The Server will also notify the Dispatcher when a new connection is made, and when that connection is closed.

This also includes a TypeScript upgrade since some of our dependencies now depend on later versions of TypeScript.