Add a package.json for the dev branch.
This is the most reliable way to get a working copy. It's not ideal since it's self-referential and doesn't include a package-lock.json file, but it *works*.

Attempting to do a global installation of the Broker resulted in only two or three JS files actually being put in the folder - most were missing and that prevented it from working. It's not clear why that was the case, even after fiddling with both the package.json and the .npmignore files to try and prevent that from happening.