fix: check if connection has been added to pool on close
Clients are only added to the connections map once they have identified themselves, however a client can close a connection before identification. And now that the token is fetched from the URL, means the if(token) check is longer an implicit "client has identified" check, so we *also* need to check if the map contains a value for the token - if it doesn't, the connection never identified itself.